Supervised Learning_Project_Thera Bank
Thera bank is interested in expanding Personal Loan customer-base rapidly to bring in more loan business and in the process, earn more through the interest on loans. A Supervised Learning Model is built to help the Bank identify the potential customers who have higher probability of purchasing the loan. This will increase the success ratio while at the same time reduce the cost of the campaign.