Supervised Learning_Patient type_Classifier
"The problem statement is to classify the type of Patient from the biomechanical features of orthopedic patients.This is a binary classification problem with two classes 'Abnormal, 'Normal'. It is a simple dataset majorly used in understanding EDA pipeline and generate good insights of data. 

A KNN classifier is used and tuned for best performance."